requirements.txt
accelerate>=1.6.0
pillow>=11.2.1
requests>=2.32.3
torch>=2.6.0
transformers>=4.51.3
